py-fixComm bugfix: wasn't announcing game to all ready hosts.
When a match was announcing a game to its chosen workers, it was
looping over all of them, but only notifying the last one. It was just
a tab problem, where the part where the socket was opened and game info
sent, was only happening *after* the loop, therefore to last PlayerHost
that was pulled from the list of the Match's hosts.