py-ctrlCOnDispatcher: ctrl+c now kills dispatcher.
Apparently, if you don't set a threading.Thread's daemon field to True,
it will not exit when the main thread does.  If you want a thread to die
when the main thread does, you just need to set t.daemon = True.
(NOTE: turns out if you do this and the main thread doesn't do anything
else after starting the thread, main will stop, and so of course the
other thread will too...  Ergo, the opposite default.
Learned about this here:
http://stackoverflow.com/questions/11815947/cannot-kill-python-script-with-ctrl-c
)