py-logging: added util.logging_help...
Because the goal is to use the logging config files, we don't get to
touch the logging handlers, because they're constructed for us.  So the
only way to get to the handlers to adjust only them is to examine the
loggers' __dict__ fields, pull out their handlers, and adjust them.
- One function sets the level of all StreamHandlers, so that we can
  adjust debug/verbose/quiet modes for the program, while leaving the
  file logs at debug level.
- One function resets the file that is used by a FileHandler, so that we
  can have log files named after the ports that are read as commandline
  arguments.