py-configFiles: made Match serliaziable, sort of
By sort of, I just mean, it's not a very standard way.
- Match.to_dict: spits out the "public" fields in a dict.
- Match.from_dict: sets (or adds!) fields from a dict
This way, we can read in a json dict from a config file to construct
Match objects.
Also, this meant changing the field that was a list of player types from
private to public, so that the file can specify what player types we
want, based on the experiment we might be running.

squash! Bugfixes

- Match.to_dict() didn't return.
- Match logging called json.dump(self) instead of self.to_dict()