dispatcher/dispatchServer: waits for players to get ready.
There's a synchronicity issue between the PlayerRunner and
GameServerRunner classes. The PlayerRunner takes longer to start up than
a GameServerRunner.  This is because the PlayerRunner uses Java
reflection to find the player class to run from a list of all classes.
And the game server will shut itself down if the players aren't ready to
play when it starts.  So the PlayerRunners processes need to be started
before the GameServerRunner.
... So for now, the DispatchServer just sits and waits for ten seconds
before starting up the GameServerRunner.