Bugfix: now saves data correctly.
ARGH.  Turns out it was just one tiny problem, that made all of my
data before be recorded wrong.  The first deployment
(tag:deploy-Minions1) recorded the computer that the player was run on,
and not what type of player it was.  Turns out, for some unknown brain
fart of a reason, when I wrote the dispatcher, I made it add hosts to
the GGPServerProcess using the hostname for both the hostname and the
playerName parameters.  Just forgot what the playerName parameter was?
I dunno.  Anyways, that was where the ggp-base code was getting the
wrong names from, not that it was somehow using hostnames as playerNames
--- I did that, but couldn't find it until I had stepped back for a
couple of weeks.